pyproject.toml
[project]
name = "user-persona-sampling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "fire>=0.7.0",
    "hf-transfer>=0.1.9",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "transformers>=4.57.3",
]
